RNP-67: Refactor AuthScreen and AuthButton to not duplicate logic
We had logic for authenticating in both the AuthButton component and in
the AuthScreen component. This logic in AuthScreen was used when
multiple profiles were added and the logic in AuthButton when a single
profile was added. The logic was also not the exact same and was
buggy for the AuthButton. This change moves all the logic from the
AuthScreen into the AuthButton component.