Revert "MVP commit"
In the past there was a huge commit made with the name "MVP". This
commit caused quite some merge conflicts and other working branches
already seem to have versions of the code that this commit contains
either in different directories or because the code was copy pasted.
So let's remove it here to avoid conflicts.

This reverts commit 7c97cbd711e2a46ee69e715d08526a08f666c327.