RNP-98: Send different events for incorrectPin and pinNotAllowed
Previously on both authentication and pin creation we would send
failures using the onError event. This is now split into incorrectPin
and pinNotAllowed. These are the only two errors that can actually occur
so we don't even need an onError event.

This commit also changes a longstanding nuisance where the
remainingFailureCount was send in a userinfo object with no additional
values. So that is now sent as it's own property.