RNP-98: iOS: change showError to onIncorrectPin and onPinNotAllowed
Similar to android, we move the combined showError events to their own
specific error events.

We here also move the remainingFailureCount out of the profile object
and send it as a direct property on the event.

We also, as on android, rename data ->  pinLength as that is what is
actually is.