RNP-91: Add support for passing authenticatorId in authenticateUser
Previously the api was exposed, so authenticateUser expected an
authenticatorId to be passed in. But it was not being used. This commit
adds that functionality by passing it into the sdk if it exists.

The developer is still responsible for checking if the passed in
authenticatorId is actually valid for the profile that he trying to
authenticate.