RNP-126: iOS: Remove actions from pin/customregistration
Previously we had an action which indicated if we wanted to supply a
pin/cancel pin and if we wanted to supply custom registration data or
cancel it. We have moved the canceling to their own api methods so we no
longer need to action to specify what we are doing and only supply the
data.