RNP-124: Refactor ResourceRequests into useCase and remove retrofit
Previously we had a lot of code for the resourceRequests using retrofit.
Since we are a wrapper we don't actually require a typesafe httpclient
as we're just passing the results back into ReactNative as a string.

Removing this allows us to fit the whole logic except for the mapper all
into the resoureRequestUseCase :^)