RNP-124: Refactor error handling + don't allow empty path
Previously we allowed users to send an empty path, this is now no longer
allowed and will reject with an error. We also no longer fallback to GET
when supplying an incorrect HTTP method as we already typecheck and
require this in typescript we should fail when we get an unexpected
method.