RNP-133: Add a makeCompletionClosure function and use it
This commit renames the completion generator into makeCompletionClosure
and uses it. We store the closure in a variable and then use that to
make it more readable.

There are a few places where we can't use this function and we have to
use the resolveResultOrRejectError method. This is because we transform
the data that we receive back so we can't use a generic function for it.