RNP-133: Simplify lots of completion closures
We were using completion closures with both a Bool and an Error? as
arguments. This is not needed as we can use the error to indicate if the
method was successful or not. These have now been removed.

We also add another makeCompletionClosure which does not take the
resolveWithNil argument as it does not even take the result argument.
It returns:  (_ error: Error?) -> Void