Exclude files from building in podspec
When linking with the RN sdk locally we get some problems in the example
app when already having built the sdk locally aswel. This happens
because we have some generated code laying around and we also copy over
all the Pods and try to compile those twice.
This results in some build errors and is fixed by just excluding these
paths in the podspec.