RNP-137: Expose all errors as enums to Typescript
We have 3 different types of errors which can occur.
1. Wrapper specific errors
2. Android errors
3. iOS errors

We expose all errors under a unified Error object. This object also
contains 3 other enums which contain the platform/wrapper specific
errors.
All of the errors here match except for 1 which can only occur in
functionality that is not supported in RN currently, custom
authenticators.
  - AuthenticatorRegistrationErrorCustomAuthenticationDisabled