RNP-144: Move login/register to swift api + fix deadlock during auth
Previously, it was possible to deadlock authentication by starting
authentication during another authentication flow.
To fix this, we need to save the completion with the authentication
delegate so we can start a new authentication and reject the correct
promise.
Due to some issues with deinitalization of the delegate with the objc
api we also convert all the login/regiter/changepin/authenticators logic
to the swift api because this api already keeps a reference to the
delegate so we don't have to do that ourselves to prevent it from
getting de-initalized.