Android: Rework config loading in RN sdk
Previously we would load the config from the startSDK function called
from RN with a config. This has now been changed to load from a class
which has to be implemented and will be loaded by reflection. Files
implementing the interface for the config will be kept in the main dex
file so it can be found at runtime.