RNP-140: Don't reject on unexpected body, instead silently finish
Previously we were rejecting if got a body in a GET request or not
receive a body in a POST request.

This commit changes that and just silently does the request without the
body for GET and with an empty body for POST/PUT.

In Typescript definitions we have defined that the body must/musn't be
present in certain resource calls.